update HTTPClient docs
